+++ src/window_utils.c
@@ -6,16 +6,16 @@
 /*   By: estettle <estettle@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/26 21:27:17 by estettle          #+#    #+#             */
-/*   Updated: 2024/11/27 11:02:27 by estettle         ###   ########.fr       */
+/*   Updated: 2024/11/27 11:27:47 by estettle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/so_long.h"
 
-void	put_pixel(t_data *data, int x, int y, long color)
+void	put_pixel(t_data *data, int x, int y, uint32_t color)
 {
 	char	*dst;
 
 	dst = data->addr + (y * data->line_length + x * (data->bits_per_pixel / 8));
-	*(unsigned int*)dst = color;
+	*(uint32_t*)dst = color;
 }
